+++ src/presenter/point-presenter.js
@@ -2,7 +2,7 @@ import EditingPoint from '../view/point-edit.js';
 import TripListLi from '../view/trip-list-li.js';
 import PointsList from '../view/points-list.js';
 import { renderPosition, render, replace, remove } from '../utils/rendering-utils.js';
-import { UserAction, UpdateType, Mode } from '../utils/common.js';
+import { UserAction, UpdateType, Mode } from '../utils/constants.js';
 
 export default class TripPoint {
   constructor(pointContainer, changeData, changeMode) {
