+++ src/utils/common.js
@@ -17,33 +17,11 @@ export const generateDate = () => {
   return dayjs().add(daysGap, 'm').toDate();
 };
 
-export const SortType = {
-  DAY: 'DAY',
-  EVENT: 'EVENT',
-  TIME: 'TIME',
-  PRICE: 'PRICE',
-  OFFERS: 'OFFERS',
+export const checkCity = (city, array) => {
+  if (city && array.some((it) => it === city)) {
+    return '';
+  }
+  return 'disabled';
 };
 
-export const UserAction = {
-  UPDATE_POINT: 'UPDATE_POINT',
-  ADD_POINT: 'ADD_POINT',
-  DELETE_POINT: 'DELETE_POINT',
-};
-
-export const UpdateType = {
-  PATCH: 'PATCH',
-  MINOR: 'MINOR',
-  MAJOR: 'MAJOR',
-};
-
-export const FilterType = {
-  EVERYTHING: 'everything',
-  FUTURE: 'future',
-  PAST: 'past',
-};
-
-export const Mode = {
-  DEFAULT: 'DEFAULT',
-  EDITING: 'EDITING',
-};
+export const checkPrice = (price) => Math.abs(price);
