+++ src/view/sort.js
@@ -1,6 +1,6 @@
 import Abstract from '../view/abstract.js';
 import { sortList, sortWords } from '../utils/rendering-data-utils.js';
-import { SortType } from '../utils/common.js';
+import { SortType } from '../utils/constants.js';
 
 const sort = (currentSortType = SortType.DAY) =>
   `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
